Reduce duplication in file and directory browsers
