+++ include/filebrowserformaction.h
@@ -14,7 +14,11 @@ namespace newsboat {
 
 class FileBrowserFormAction : public FormAction {
 public:
-	FileBrowserFormAction(View&, std::string formstr, ConfigContainer* cfg);
+	enum class Variant {
+		FileSelection,
+		DirectorySelection,
+	};
+	FileBrowserFormAction(View&, std::string formstr, ConfigContainer* cfg, Variant variant);
 	~FileBrowserFormAction() override = default;
 	void prepare() override;
 	void init() override;
@@ -27,6 +31,12 @@ public:
 
 	Dialog id() const override
 	{
+		switch (variant) {
+		case Variant::FileSelection:
+			return Dialog::FileBrowser;
+		case Variant::DirectorySelection:
+			return Dialog::DirBrowser;
+		}
 		return Dialog::FileBrowser;
 	}
 	std::string title() override;
@@ -34,7 +44,13 @@ public:
 protected:
 	std::string main_widget() const override
 	{
-		return "filename";
+		switch (variant) {
+		case Variant::FileSelection:
+			return "filename";
+		case Variant::DirectorySelection:
+			return "files";
+		}
+		return "";
 	}
 
 private:
@@ -49,6 +65,8 @@ private:
 	std::vector<StflRichText> lines;
 
 	std::string get_formatted_filename(const Filepath& filename, mode_t mode);
+
+	Variant variant;
 
 	LineView file_prompt_line;
 	Filepath default_filename;
